framework/requirements.txt
Appium-Python-Client==1.0.2
attrs==20.3.0
bottle==0.12.19
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
pluralize==0.1.6
py==1.9.0
py4web==1.20201024.1
pydal==20200910.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
requests==2.25.0
selenium==3.141.0
six==1.15.0
threadsafevariable==1.0
toml==0.10.2
tornado==6.1
urllib3==1.26.1
watchgod==0.6
yatl==20200711.1
zope.event==4.5.0
zope.interface==5.2.0
